fix: search

+++ demo/gdms-search/src/main/java/org/graduationdesign/search/entity/Message.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Component;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Message {
-    private int id;
-    private int type;
+    private double id;
+    private double type;
     private String title;
     private String content;
 }
